CodePage/requirements.txt
requests
six
django_wysiwyg
django_ace